+++ src/shared/var_main_mod.f90
@@ -8,7 +8,7 @@ module var_main
 
   character(len=150) :: MEASURE_ADJ_OUTDIR
   
-  real :: MIN_PERIOD, MAX_PERIOD
+  double precision :: MIN_PERIOD, MAX_PERIOD
 
   integer :: weighting_option
